backend/requirements.txt
fastapi
uvicorn[standard]
gtts
deep-translator
pytesseract
Pillow
python-multipart
requests
fasttext-wheel
indic-transliteration
transformers
torch
